+++ src/base/DataSet.java
@@ -4,14 +4,14 @@ package base;
  * Created by galafit on 10/10/17.
  */
 public interface DataSet {
-    public float getYValue(int index, int yColumnNumber);
+    public double getYValue(int index, int yColumnNumber);
 
     /**
      * amount of Y columns
      */
     public int getYColumnsCounter();
 
-    public float getXValue(int index);
+    public double getXValue(int index);
 
     public String getAnnotation(int index);
 
@@ -21,8 +21,8 @@ public interface DataSet {
 
     public Range getYExtremes(int yColumnNumber);
 
-    public int findNearestData(float xValue);
+    public int findNearestData(double xValue);
 
-    public float getAverageDataInterval();
+    public double getAverageDataInterval();
 
 }
