+++ src/base/XYViewer.java
@@ -14,11 +14,11 @@ public class XYViewer {
         return dataSet.size();
     }
 
-    public float getX(int index) {
+    public double getX(int index) {
         return dataSet.getXValue(index);
     }
 
-    public float getY(int index) {
+    public double getY(int index) {
         return dataSet.getYValue(index, 0);
     }
 
@@ -30,7 +30,7 @@ public class XYViewer {
         return dataSet.getXExtremes();
     }
 
-    public int findNearest(float xValue) {
+    public int findNearest(double xValue) {
         return dataSet.findNearestData(xValue);
     }
